+++ bpwave/__init__.py
@@ -1,3 +1,4 @@
+from . import visu
 from ._version import __version__
 from .io import CsvReader, SignalReader, to_csv
 from .signal import ChPoints, CpIndices, Signal